re-instate important check lost in 016a075e79

+++ nouveau/lucene9/src/main/java/org/apache/couchdb/nouveau/lucene9/core/Lucene9Index.java
@@ -136,9 +136,11 @@ public class Lucene9Index extends Index<IndexableField> {
                     writer.rollback();
                 },
                 () -> {
-                    var dir = writer.getDirectory();
-                    for (final String name : dir.listAll()) {
-                        dir.deleteFile(name);
+                    if (isDeleteOnClose()) {
+                        var dir = writer.getDirectory();
+                        for (final String name : dir.listAll()) {
+                            dir.deleteFile(name);
+                        }
                     }
                 });
     }
